Include peername in timeout responses (most of the time) (#13)
#11 made it so that the peername info comes from kadabra in the message
response. However, on timeout, we don't _have_ a response to get this
from, so peername is blank, but timeouts are the ones where you'd really
like to see the peername.

This isn't a great solution, but it's simple: whenever we get a
response, we store the peername in the associated worker. For timeouts,
we use the value stored in the worker, which represents the last-seen
peername. That means we'll still get nil peernames if a timeout occurs
before any reponse has been received, but hopefully that will be very
rare.

I also synchronized `mix.lock` with `mix.exs` which was missed in #12.